Fix project urls in pyproject.toml
I noticed that the project urls "homepage" and "repository" were pointing to mne-bids-pipeline and not mne-qt-browser.
I don't think there is a homepage, right? An option might be to just remove the homepage url if there isn't any. I made it point to the GitHub repo for the moment.
Additionally changed "changelog" to https url.
Sorry for not creating an issue first, I was a bit lazy and I was hoping this would be okay since there is no change to the actual code.